requirements.txt
# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Gmail API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0

# Data
pandas>=2.1.0
plotly>=5.18.0

# HTTP (used by unsubscribe_via_post)
requests>=2.31.0

# Dev / testing
pytest>=8.0.0
httpx>=0.27.0
